DASHB-660 : Generation SESSION-ID for analytics changed.
